JS-2024 Refactor S6551 helper logic
